pdf_chatbot_functional_chat_ui/requirements.txt
streamlit
langchain
openai
faiss-cpu
PyMuPDF
tiktoken
python-dotenv
